# medienverwaltungweb.client/swagger.json
{
    "openapi": "3.0.4",
    "info": {
      "title": "medienVerwaltungWeb.Server",
      "version": "1.0"
    },
    "paths": {
...
      "/api/Song": {
        "get": {
          "tags": [
            "Song"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SongDTO"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SongDTO"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SongDTO"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Song"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongDTO"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/SongDTO"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SongDTO"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SongDTO"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SongDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Song/GetSongsByCount/{count}": {
        "get": {
          "tags": [
            "Song"
          ],
          "parameters": [
            {
              "name": "count",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SongDTO"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SongDTO"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SongDTO"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/Song/{id}": {
        "get": {
          "tags": [
            "Song"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SongDTO"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SongDTO"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SongDTO"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Song"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongDTO"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/SongDTO"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "delete": {
          "tags": [
            "Song"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "SongDTO": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "title": {
              "type": "string",
              "nullable": true
            },
            "interpretFullName": {
              "type": "string",
              "nullable": true
            },
            "location": {
              "type": "string",
              "nullable": true
            },
            "length": {
              "type": "integer",
              "format": "int32"
            }
          },
          "additionalProperties": false
        }
      }
    }
  }